backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Celery & Redis
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# NLP Stack
spacy==3.7.2
vaderSentiment==3.3.2
transformers==4.36.2
torch==2.1.2
langdetect==1.0.9
nltk==3.8.1

# OCR
easyocr==1.7.1
Pillow==10.2.0

# HTTP Client (for scraping)
httpx==0.26.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0

# Search
meilisearch==0.28.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
structlog==24.1.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
